+++ packages/components-angular/src/app/pages/divider.component.ts
@@ -2,19 +2,23 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'page-divider',
-  styles: [`
-    .divider-vertical-container-example {
-      display: flex;
-      height: 100px;
-    }
+  styles: [
+    `
+      .divider-vertical-container-example {
+        display: flex;
+        height: 100px;
+      }
 
-    /*@media (min-width: 480px) and (max-width: 759px), (min-width: 1000px) and (max-width: 1299px), (min-width: 1760px) {*/
-    /*  .divider-vertical-responsive-container-example {*/
-    /*    display: flex;*/
-    /*    height: 100px;*/
-    /*  }*/
-    /*}*/
-  `],
+      @media (min-width: 480px) and (max-width: 759px),
+        (min-width: 1000px) and (max-width: 1299px),
+        (min-width: 1760px) {
+        .divider-vertical-responsive-container-example {
+          display: flex;
+          height: 100px;
+        }
+      }
+    `
+  ],
   template: `
     <div class="playground light" title="should show a divider">
       <p-divider></p-divider>
@@ -26,12 +30,13 @@ import { Component } from '@angular/core';
       </div>
     </div>
 
-<!--    <div class="playground light" title="should show responsive vertical divider">-->
-<!--      <div class="divider-vertical-responsive-container-example">-->
-<!--        <p-divider-->
-<!--          orientation="{base: 'horizontal', xs: 'vertical', s: 'horizontal', m: 'vertical', l: 'horizontal', xl: 'vertical'}"></p-divider>-->
-<!--      </div>-->
-<!--    </div>-->
+    <div class="playground light" title="should show responsive vertical divider">
+      <div class="divider-vertical-responsive-container-example">
+        <p-divider
+          orientation="{base: 'horizontal', xs: 'vertical', s: 'horizontal', m: 'vertical', l: 'horizontal', xl: 'vertical'}"
+        ></p-divider>
+      </div>
+    </div>
 
     <div class="playground light" title="should show different colors of divider with light theme">
       <p-divider></p-divider>
@@ -54,5 +59,4 @@ import { Component } from '@angular/core';
     </div>
   `
 })
-export class DividerComponent {
-}
+export class DividerComponent {}