+++ packages/components-react/src/pages/Divider.tsx
@@ -29,11 +29,11 @@ export const DividerPage = (): JSX.Element => {
         </div>
       </div>
 
-      {/*<div className="playground light" title="should show responsive vertical divider">*/}
-      {/*  <div className="divider-vertical-responsive-container-example">*/}
-      {/*    <Divider orientation="{base: 'horizontal', xs: 'vertical', s: 'horizontal', m: 'vertical', l: 'horizontal', xl: 'vertical'}" />*/}
-      {/*  </div>*/}
-      {/*</div>*/}
+      <div className="playground light" title="should show responsive vertical divider">
+        <div className="divider-vertical-responsive-container-example">
+          <Divider orientation="{base: 'horizontal', xs: 'vertical', s: 'horizontal', m: 'vertical', l: 'horizontal', xl: 'vertical'}" />
+        </div>
+      </div>
 
       <div className="playground light" title="should show different colors of divider with light theme">
         <Divider />