+++ packages/components/src/components/layout/divider/divider.tsx
@@ -1,6 +1,6 @@
 import { Component, h, JSX, Prop } from '@stencil/core';
 import cx from 'classnames';
-import { mapBreakpointPropToPrefixedClasses, prefix } from '../../../utils';
+import { BreakpointCustomizable, mapBreakpointPropToPrefixedClasses, prefix } from '../../../utils';
 import { Theme } from '../../../types';
 
 @Component({
@@ -10,10 +10,11 @@ import { Theme } from '../../../types';
 })
 export class Divider {
   /** Defines color depending on theme. */
-  @Prop() public color?: 'neutral-contrast-high' | 'neutral-contrast-medium' | 'neutral-contrast-low' = 'neutral-contrast-low';
+  @Prop() public color?: 'neutral-contrast-high' | 'neutral-contrast-medium' | 'neutral-contrast-low' =
+    'neutral-contrast-low';
 
   /** Defines orientation. */
-  @Prop() public orientation?: 'vertical' | 'horizontal' = 'horizontal';
+  @Prop() public orientation?: BreakpointCustomizable<'vertical' | 'horizontal'> = 'horizontal';
 
   /** Adapts color depending on theme. */
   @Prop() public theme?: Theme = 'light';
@@ -26,8 +27,6 @@ export class Divider {
       mapBreakpointPropToPrefixedClasses('divider--orientation', this.orientation)
     );
 
-    return (
-      <hr class={dividerClasses}/>
-    );
+    return <hr class={dividerClasses} />;
   }
 }