+++ packages/components/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Fixed
 - Mix of `Optgroups` and `Options` on same level in `Select Wrapper` component
+- Fix typings for `orientation` of `Divider` component
 
 ### [1.5.2] - 2020-07-22